pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf2foundry"
version = "0.1.0"
description = "Convert born-digital PDFs into Foundry VTT v13 module compendia."
readme = "README.md"
requires-python = ">=3.12"
authors = [
  { name = "Martin Papy", email = "martin.papy@gmail.com" }
]
dependencies = [
  "typer>=0.16.0",
  "jinja2>=3.1.6",
  "pillow>=11.3.0",
  "rich>=14.1.0",
  "docling>=2.53.0",
  "docling-core>=2.48.1",
  "pytesseract>=0.3.13",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.13.1",
  "black>=25.9.0",
  "mypy>=1.18.2",
  "pytest>=8.4.2",
  "pytest-cov>=7.0.0",
  "pre-commit>=4.3.0",
]

[project.urls]
Homepage = "https://github.com/martin-papy/pdf2foundry"

[project.scripts]
pdf2foundry = "pdf2foundry.cli:app"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
  "E","F","I","B","UP","C4","SIM","RUF"
]
ignore = ["ISC001"]

[tool.ruff.lint.isort]
known-first-party = ["pdf2foundry"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true
pretty = true
exclude = [
  "^tests/.*$"
]
# Performance optimizations
sqlite_cache = true
cache_fine_grained = true
incremental = true
# Skip checking the contents of functions with at least one `Any` parameter
disallow_any_generics = false
# Faster imports
follow_imports = "normal"
# Reduce memory usage
local_partial_types = true

[[tool.mypy.overrides]]
module = [
  "docling",
  "docling.*",
  "docling_core",
  "docling_core.*",
  "pytest",
  "jinja2",
  "pytesseract",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-q --cov=pdf2foundry --cov-report=term-missing --cov-fail-under=90"
testpaths = ["tests"]

